fix for revision 832581 as reported in the mailing list and waiting for adirans attention
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@833307 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/workeffort/src/org/ofbiz/workeffort/workeffort/WorkEffortServices.java
@@ -637,7 +637,7 @@ public class WorkEffortServices {
                             for (DateRange periodRange : periodRanges) {
                                 if (periodRange.includesDate(occurrence)) {
                                     GenericValue cloneWorkEffort = (GenericValue) workEffort.clone();
-                                    TimeDuration duration = TimeDuration.fromNumber(workEffort.getLong("estimatedMilliSeconds"));
+                                    TimeDuration duration = TimeDuration.fromNumber(workEffort.getDouble("estimatedMilliSeconds").longValue());
                                     if (!duration.isZero()) {
                                         Calendar endCal = UtilDateTime.toCalendar(occurrence, timeZone, locale);
                                         Date endDate = duration.addToCalendar(endCal).getTime();
